+++ terraform/provider/main.go
@@ -1,0 +1,18 @@
+package main
+
+import (
+	"github.com/chanzuckerberg/happy/terraform/provider/pkg/provider"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/plugin"
+)
+
+// Generate the Terraform provider documentation using `tfplugindocs`:
+//go:generate go run github.com/hashicorp/terraform-plugin-docs/cmd/tfplugindocs generate --provider-name happy
+
+func main() {
+	plugin.Serve(&plugin.ServeOpts{
+		ProviderFunc: func() *schema.Provider {
+			return provider.Provider()
+		},
+	})
+}