+++ api/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## [0.39.0](https://github.com/chanzuckerberg/happy/compare/api-v0.38.0...api-v0.39.0) (2022-10-25)
+
+
+### Features
+
+* create terraform provider for happy-api ([#699](https://github.com/chanzuckerberg/happy/issues/699)) ([3325039](https://github.com/chanzuckerberg/happy/commit/3325039ae0fa433ee4d59307762869ed543b8554))
+
+## [0.38.0](https://github.com/chanzuckerberg/happy/compare/api-v0.37.0...api-v0.38.0) (2022-10-24)
+
+
+### Features
+
+* disable auth for test and dev environments ([#672](https://github.com/chanzuckerberg/happy/issues/672)) ([6c4000f](https://github.com/chanzuckerberg/happy/commit/6c4000fcbfff2169741d3bbe6c1d3366c71b204d))
+* roll out config feature to cli ([#660](https://github.com/chanzuckerberg/happy/issues/660)) ([a72c965](https://github.com/chanzuckerberg/happy/commit/a72c965f6bd2c9113c8152c9155330971e808b46))
+
 ## [0.37.0](https://github.com/chanzuckerberg/happy/compare/api-v0.36.0...api-v0.37.0) (2022-10-13)